Move some get/set lyric property from singer manager to lyric manager.
