+++ osu.Game.Rulesets.Karaoke/Edit/Singers/Components/Timeline/LyricTimelineSelectionHandler.cs
@@ -6,6 +6,8 @@ using System.Linq;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics.UserInterface;
 using osu.Game.Rulesets.Edit;
+using osu.Game.Rulesets.Karaoke.Edit.Components.ContextMenu;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Screens.Edit.Compose.Components;
 using static osu.Game.Rulesets.Karaoke.Edit.Components.Timeline.TimelineBlueprintContainer;
@@ -15,11 +17,15 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Singers.Components.Timeline
     internal class LyricTimelineSelectionHandler : TimelineSelectionHandler
     {
         [Resolved]
-        private SingerManager singerManager { get; set; }
+        private LyricManager lyricManager { get; set; }
 
         public override bool HandleMovement(MoveSelectionEvent moveEvent) => false;
 
         protected override IEnumerable<MenuItem> GetContextMenuItemsForSelection(IEnumerable<SelectionBlueprint> selection)
-            => singerManager.CreateSingerContextMenu(selection.Select(x => x.HitObject).OfType<Lyric>().ToList());
+        {
+            var lyrics = selection.Select(x => x.HitObject).OfType<Lyric>().ToList();
+            var contectMenu = new SingerContextMenu(lyricManager, lyrics, "");
+            return contectMenu.Items;
+        }
     }
 }
